# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reports:
    get:
      operationId: listReports
      tags: [reports]
      summary: List user's reports
      responses:
        "200":
          description: List of reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReportSummary"
    post:
      operationId: createReport
      tags: [reports]
      summary: Save a new report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReportBody"
      responses:
        "201":
          description: Created report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportCreated"
  /reports/{id}:
    get:
      operationId: getReport
      tags: [reports]
      summary: Get a report by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportRow"
    delete:
      operationId: deleteReport
      tags: [reports]
      summary: Delete a report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /reports/{id}/status:
    patch:
      operationId: updateReportStatus
      tags: [reports]
      summary: Update report status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateReportStatusBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /reports/slug/{slug}:
    get:
      operationId: getReportBySlug
      tags: [reports]
      summary: Get a report by slug (public)
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportRow"
  /reports/{id}/output:
    get:
      operationId: getReportOutput
      tags: [reports]
      summary: Get report output only
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Report output
          content:
            application/json:
              schema:
                type: object
...
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ReportSummary:
      type: object
      properties:
        id:
          type: string
        slug:
          type: string
        title:
          type: string
        industry:
          type: string
          nullable: true
        status:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - slug
        - title
        - status
        - createdAt
        - updatedAt
    ReportRow:
      type: object
      properties:
        id:
          type: string
        slug:
          type: string
        userId:
          type: string
        title:
          type: string
        industry:
          type: string
          nullable: true
        inputs:
          type: object
        output:
          type: object
        status:
          type: string
        isPublic:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - slug
        - userId
        - title
        - inputs
        - output
        - status
        - isPublic
        - createdAt
        - updatedAt
    ReportCreated:
      type: object
      properties:
        id:
          type: string
        slug:
          type: string
      required:
        - id
        - slug
    CreateReportBody:
      type: object
      properties:
        title:
          type: string
        industry:
          type: string
          nullable: true
        inputs:
          type: object
        output:
          type: object
      required:
        - title
        - inputs
        - output
    UpdateReportStatusBody:
      type: object
      properties:
        status:
          type: string
        note:
          type: string
          nullable: true
      required:
        - status
...