# lib/api-zod/src/generated/types/reportRow.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ReportRowInputs } from "./reportRowInputs";
import type { ReportRowOutput } from "./reportRowOutput";

export interface ReportRow {
  id: string;
  slug: string;
  userId: string;
  title: string;
  industry?: string | null;
  inputs: ReportRowInputs;
  output: ReportRowOutput;
  status: string;
  isPublic: boolean;
  createdAt: string;
  updatedAt: string;
}